+++ crates/p2p/src/lib.rs
@@ -3,3 +3,6 @@
 pub mod commands;
 pub mod events;
 pub mod swarm;
+
+#[cfg(test)]
+mod tests;
