+++ crates/p2p/src/tests/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod common;
+pub(crate) mod gossipsub;
+pub(crate) mod request;
